Added a new Maven plugin configuration parameter named 'runOnlyAtExecutionRoot' to allow for the plugin to execute in each module of a multi-module reactor, or restrict to running once in a reactor.
The default value is 'true' which maintains the existing behavior of
the plugin to run only once in a reactor build.

The current plugin version (1.2.6) only supports single execution in a
multi-module reactor, but its logic assumes and therefore requires
a certain Maven project layout which is not suitable for all.

This change simply allows for this plugin to be used in Maven
multi-module projects without error if they do not conform to the
expected project layout, at the cost of having the information extracted
for each module.

The false assumption necessitating this enhancement is that the root
pom.xml in a reactor/multi-module project is also the parent pom.xml for
all modules within that reactor. That is a common Maven project setup, but
it is not a mandatory one. The parent pom.xml can exist as a submodule
or outside of the reactor entirely.

In my case, my root/reactor pom.xml which designates the module list is
not the parent pom.xml. A parent pom.xml is the first module in the
multi-module project. The root/reactor pom.xml simply designates the
module list and handles a final package assembly step as the last part
of the reactor build. Without the ability to disable the
runOnlyAtExecutionRoot behavior, the maven-jgit-buildnumber-plugin
fails and is unusable.